requirements.txt
tzlocal==5.3.1
python-dotenv==1.2.1
openai==2.8.1
requests==2.32.5
yt-dlp==2025.11.12